backend/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Jinja2==3.1.2

# Authentication and security
bcrypt==3.2.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Form handling and validation
python-multipart==0.0.6
email-validator==2.1.0

# Database utilities
sqlalchemy-utils==0.41.1

# Environment and configuration
python-dotenv==1.0.0

# Production dependencies
gunicorn==21.2.0
uvloop==0.19.0
httptools==0.6.1